[Spark] move spark.sql classes to paimon package
